init commit of refactoring roles (probably broken)
