+++ packages/agent/src/swarm/communityManager/index.ts
@@ -1,14 +1,8 @@
-// Community manager
-
-// The community manager greets new users and helps them get started
-// The community manager also helps moderators with moderation tasks, including banning scammers
+import type { Character, IAgentRuntime, OnboardingConfig } from "@elizaos/core";
+import type { Guild } from 'discord.js';
 import dotenv from "dotenv";
-dotenv.config({ path: "../../.env" });
-
-import type { Character, IAgentRuntime } from "@elizaos/core";
-import type { Guild } from "discord.js";
-import { initializeAllSystems } from "../shared/onboarding/initialize";
-import type { OnboardingConfig } from "../shared/onboarding/types";
+import { initializeAllSystems } from "../settings";
+dotenv.config({ path: '../../.env' });
 
 const character: Character = {
   name: "Kelsey",
