+++ packages/agent/src/swarm/counselor/index.ts
@@ -1,10 +1,8 @@
+import type { Character, IAgentRuntime, OnboardingConfig } from "@elizaos/core";
+import type { Guild } from 'discord.js';
 import dotenv from "dotenv";
-dotenv.config({ path: "../../.env" });
-
-import type { Character, IAgentRuntime } from "@elizaos/core";
-import type { Guild } from "discord.js";
-import { initializeAllSystems } from "../shared/onboarding/initialize";
-import type { OnboardingConfig } from "../shared/onboarding/types";
+import { initializeAllSystems } from "../settings";
+dotenv.config({ path: '../../.env' });
 
 const character: Character = {
   name: "Ruby",
