--- packages/core/src/actions/index.ts
@@ -1,8 +1,0 @@
-export * from "./followRoom.ts";
-export * from "./ignore.ts";
-export * from "./muteRoom.ts";
-export * from "./none.ts";
-export * from "./unfollowRoom.ts";
-export * from "./unmuteRoom.ts";
-export * from "./confirm.ts";
-export * from "./cancel.ts";
