+++ packages/core/src/index.ts
@@ -10,6 +10,7 @@ export * from "./environment.ts";
 export * from "./evaluators.ts";
 export * from "./generation.ts";
 export * from "./goals.ts";
+export * from "./import.ts";
 export * from "./knowledge.ts";
 export * from "./logger.ts";
 export * from "./memory.ts";
@@ -18,7 +19,7 @@ export * from "./parsing.ts";
 export * from "./posts.ts";
 export * from "./providers.ts";
 export * from "./relationships.ts";
+export * from "./roles.ts";
 export * from "./runtime.ts";
 export * from "./settings.ts";
 export * from "./uuid.ts";
-export * from "./import.ts";
