--- packages/core/src/providers/index.ts
@@ -1,2 +1,0 @@
-export * from "./time.ts";
-export * from "./facts.ts";
