+++ packages/core/src/settings.ts
@@ -1,144 +1,120 @@
-import { config } from "dotenv";
-import fs from "node:fs";
-import path from "node:path";
-import logger from "./logger.ts";
+import { logger } from "./logger";
+import { OnboardingSetting, IAgentRuntime, OnboardingState, OnboardingConfig, WorldData } from "./types";
+import { stringToUuid } from "./uuid";
 
-interface Settings {
-  [key: string]: string | undefined;
+function createSettingFromConfig(
+  configSetting: Omit<OnboardingSetting, "value">
+): OnboardingSetting {
+  return {
+    name: configSetting.name,
+    description: configSetting.description,
+    usageDescription: configSetting.usageDescription || "",
+    value: null,
+    required: configSetting.required,
+    validation: configSetting.validation || null,
+    public: configSetting.public || false,
+    secret: configSetting.secret || false,
+    dependsOn: configSetting.dependsOn || [],
+    onSetAction: configSetting.onSetAction || null,
+    visibleIf: configSetting.visibleIf || null,
+  };
 }
 
-interface NamespacedSettings {
-  [namespace: string]: Settings;
-}
-
-let environmentSettings: Settings = {};
-
 /**
- * Determines if code is running in a browser environment
- * @returns {boolean} True if in browser environment
+ * Updates onboarding state in world metadata
  */
-const isBrowser = (): boolean => {
-  return (
-    typeof window !== "undefined" && typeof window.document !== "undefined"
-  );
-};
+export async function updateOnboardingState(
+  runtime: IAgentRuntime,
+  serverId: string,
+  onboardingState: OnboardingState
+): Promise<boolean> {
+  try {
+    const worldId = stringToUuid(`${serverId}-${runtime.agentId}`);
+    const world = await runtime.getWorld(worldId);
 
-/**
- * Recursively searches for a .env file starting from the current directory
- * and moving up through parent directories (Node.js only)
- * @param {string} [startDir=process.cwd()] - Starting directory for the search
- * @returns {string|null} Path to the nearest .env file or null if not found
- */
-export function findNearestEnvFile(startDir = process.cwd()) {
-  if (isBrowser()) return null;
-
-  let currentDir = startDir;
-
-  // Continue searching until we reach the root directory
-  while (currentDir !== path.parse(currentDir).root) {
-    const envPath = path.join(currentDir, ".env");
-
-    if (fs.existsSync(envPath)) {
-      return envPath;
+    if (!world) {
+      logger.error(`No world found for server ${serverId}`);
+      return false;
     }
 
-    // Move up to parent directory
-    currentDir = path.dirname(currentDir);
-  }
+    // Initialize metadata if it doesn't exist
+    if (!world.metadata) {
+      world.metadata = {};
+    }
 
-  // Check root directory as well
-  const rootEnvPath = path.join(path.parse(currentDir).root, ".env");
-  return fs.existsSync(rootEnvPath) ? rootEnvPath : null;
+    // Update onboarding state
+    world.metadata.onboarding = onboardingState;
+
+    // Save updated world
+    await runtime.databaseAdapter.updateWorld(world, runtime.agentId);
+
+    return true;
+  } catch (error) {
+    logger.error(`Error updating onboarding state: ${error}`);
+    return false;
+  }
 }
 
 /**
- * Configures environment settings for browser usage
- * @param {Settings} settings - Object containing environment variables
+ * Gets onboarding state from world metadata
  */
-export function configureSettings(settings: Settings) {
-  environmentSettings = { ...settings };
+export async function getOnboardingState(
+  runtime: IAgentRuntime,
+  serverId: string
+): Promise<OnboardingState | null> {
+  try {
+    const worldId = stringToUuid(`${serverId}-${runtime.agentId}`);
+    const world = await runtime.getWorld(worldId);
+
+    if (!world || !world.metadata?.onboarding) {
+      return null;
+    }
+
+    return world.metadata.onboarding as OnboardingState;
+  } catch (error) {
+    logger.error(`Error getting onboarding state: ${error}`);
+    return null;
+  }
 }
 
 /**
- * Loads environment variables from the nearest .env file in Node.js
- * or returns configured settings in browser
- * @returns {Settings} Environment variables object
- * @throws {Error} If no .env file is found in Node.js environment
+ * Initializes onboarding configuration for a server
  */
-export function loadEnvConfig(): Settings {
-  // For browser environments, return the configured settings
-  if (isBrowser()) {
-    return environmentSettings;
+export async function initializeOnboardingConfig(
+  runtime: IAgentRuntime,
+  world: WorldData,
+  config: OnboardingConfig
+): Promise<OnboardingState | null> {
+  try {
+    // Check if onboarding state already exists
+    if (world.metadata?.onboarding) {
+      logger.info(`Onboarding state already exists for server ${world.serverId}`);
+      return world.metadata.onboarding as OnboardingState;
+    }
+    
+    // Create new onboarding state
+    const onboardingState: OnboardingState = {};
+    
+    // Initialize settings from config
+    if (config.settings) {
+      for (const [key, configSetting] of Object.entries(config.settings)) {
+        onboardingState[key] = createSettingFromConfig(configSetting);
+      }
+    }
+    
+    // Save onboarding state to world metadata
+    if (!world.metadata) {
+      world.metadata = {};
+    }
+    
+    world.metadata.onboarding = onboardingState;
+    
+    await runtime.databaseAdapter.updateWorld(world, runtime.agentId);
+    
+    logger.info(`Initialized onboarding config for server ${world.serverId}`);
+    return onboardingState;
+  } catch (error) {
+    logger.error(`Error initializing onboarding config: ${error}`);
+    return null;
   }
-
-  // Node.js environment: load from .env file
-  const envPath = findNearestEnvFile();
-
-  // attempt to Load the .env file into process.env
-  const result = config(envPath ? { path: envPath } : {});
-
-  if (!result.error) {
-    logger.log(`Loaded .env file from: ${envPath}`);
-  }
-
-  // Parse namespaced settings
-  const namespacedSettings = parseNamespacedSettings(process.env as Settings);
-
-  // Attach to process.env for backward compatibility
-  Object.entries(namespacedSettings).forEach(([namespace, settings]) => {
-    process.env[`__namespaced_${namespace}`] = JSON.stringify(settings);
-  });
-
-  return process.env as Settings;
 }
-
-/**
- * Gets a specific environment variable
- * @param {string} key - The environment variable key
- * @param {string} [defaultValue] - Optional default value if key doesn't exist
- * @returns {string|undefined} The environment variable value or default value
- */
-export function getEnvVariable(
-  key: string,
-  defaultValue?: string
-): string | undefined {
-  if (isBrowser()) {
-    return environmentSettings[key] || defaultValue;
-  }
-  return process.env[key] || defaultValue;
-}
-
-/**
- * Checks if a specific environment variable exists
- * @param {string} key - The environment variable key
- * @returns {boolean} True if the environment variable exists
- */
-export function hasEnvVariable(key: string): boolean {
-  if (isBrowser()) {
-    return key in environmentSettings;
-  }
-  return key in process.env;
-}
-
-// Add this function to parse namespaced settings
-function parseNamespacedSettings(env: Settings): NamespacedSettings {
-  const namespaced: NamespacedSettings = {};
-
-  for (const [key, value] of Object.entries(env)) {
-    if (!value) continue;
-
-    const [namespace, ...rest] = key.split(".");
-    if (!namespace || rest.length === 0) continue;
-
-    const settingKey = rest.join(".");
-    namespaced[namespace] = namespaced[namespace] || {};
-    namespaced[namespace][settingKey] = value;
-  }
-
-  return namespaced;
-}
-
-// Initialize settings based on environment
-export const settings = isBrowser() ? environmentSettings : loadEnvConfig();
-
-export default settings;
